// app/lib/types.ts
export interface Page {
// ...
  servingAreasSection?: {
    enabled: boolean;
    title?: any;
    description?: any;
    serviceSlug?: string;
    serviceIds?: string[];
    areas?: unknown[];
  };
// ...
}